log: corrected currentTime property
- instead of now tracking the current time, it tracks the ratio of time
  passed since last 12th hour.

+++ src/qml/nodemodel.h
@@ -29,7 +29,7 @@ class NodeModel : public QObject
     Q_PROPERTY(int remainingSyncTime READ remainingSyncTime NOTIFY remainingTimeChanged)
     Q_PROPERTY(double verificationProgress READ verificationProgress NOTIFY verificationProgressChanged)
     Q_PROPERTY(QVariantList blockTimeList READ blockTimeList NOTIFY blockTimeListChanged)
-    Q_PROPERTY(int currentTime READ currentTime NOTIFY currentTimeChanged)
+    Q_PROPERTY(double currentTime READ currentTime NOTIFY currentTimeChanged)
 
 public:
     explicit NodeModel(interfaces::Node& node);
@@ -41,7 +41,7 @@ public:
     void setVerificationProgress(double new_progress);
     QVariantList blockTimeList() const { return m_block_time_list; }
     void setBlockTimeList(int new_block_time);
-    int currentTime() const { return m_current_time; }
+    double currentTime() const { return m_current_time; }
 
     Q_INVOKABLE void startNodeInitializionThread();
 
@@ -69,7 +69,7 @@ private:
     int m_remaining_time{0};
     double m_verification_progress{0.0};
     QVariantList m_block_time_list{0};
-    int m_current_time{0};
+    double m_current_time{0.0};
 
     int m_shutdown_polling_timer_id{0};
